Separate collision management methods from RosClient into MoveIt planner
