+++ src/compas_fab/backends/ros/planner_backend.py
@@ -10,6 +10,7 @@ from __future__ import print_function
 
 from roslibpy import Service
 from roslibpy import ServiceRequest
+
 from compas_fab.backends.ros.exceptions import RosValidationError
 
 __all__ = [
@@ -23,6 +24,10 @@ class PlannerBackend(object):
     def validate_response(self, response):
         pass
 
+    # ==========================================================================
+    # planning services
+    # ==========================================================================
+
     def inverse_kinematics_async(self, *args, **kwargs):
         pass
 
@@ -33,6 +38,25 @@ class PlannerBackend(object):
         pass
 
     def plan_cartesian_motion_async(self, *args, **kwargs):
+        pass
+
+    # ==========================================================================
+    # collision objects
+    # ==========================================================================
+
+    def add_collision_mesh(self, collision_mesh):
+        pass
+
+    def remove_collision_mesh(self, id):
+        pass
+
+    def append_collision_mesh(self, collision_mesh):
+        pass
+
+    def add_attached_collision_mesh(self, attached_collision_mesh):
+        pass
+
+    def remove_attached_collision_mesh(self, id):
         pass
